# apps/api/app/schemas/reaction.py
from marshmallow import Schema, fields, validate
 
from ..models.reaction import REACTION_KINDS, REACTION_TARGET_TYPES
 
 
class ReactionSchema(Schema):
    id = fields.Int(dump_only=True)
    # 所有者は current_user.id からサーバー側で補う(憲章 原則III)。クライアント指定は受け付けない。
    user_id = fields.Int(dump_only=True)
    target_type = fields.Str(required=True, validate=validate.OneOf(REACTION_TARGET_TYPES))
    target_id = fields.Int(required=True)
    kind = fields.Str(required=True, validate=validate.OneOf(REACTION_KINDS))
    created_at = fields.DateTime(dump_only=True)
 
 
reaction_schema = ReactionSchema()
reactions_schema = ReactionSchema(many=True)