# openapi.json
{
  "info": {
    "description": "docs/design.md §7 の契約から生成",
    "title": "hackstage api",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/reactions": {
      "get": {
        "summary": "対象(投稿/予定)のリアクション一覧を取得する",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "正常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Reaction"
                  }
                }
              }
            }
          },
          "400": {
            "description": "target_type・target_idが不正"
          },
          "401": {
            "description": "未ログイン"
          }
        }
      },
      "post": {
        "summary": "リアクションを追加する。user_idはクライアントから指定できない",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reaction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "作成成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reaction"
                }
              }
            }
          },
          "200": {
            "description": "既存のリアクションのkindを上書き",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reaction"
                }
              }
            }
          },
          "400": {
            "description": "入力エラー"
          },
          "401": {
            "description": "未ログイン"
          },
          "404": {
            "description": "対象が存在しない"
          }
        }
      }
    },
    "/api/reactions/{reaction_id}": {
      "delete": {
        "summary": "自分のリアクションを削除する。他人のリアクションは404",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "削除成功"
          },
          "401": {
            "description": "未ログイン"
          },
          "404": {
            "description": "存在しない、または他人のリアクション"
          }
        }
      }
    },
    "/api/calendars/mine": {
      "get": {
        "summary": "自分の個人カレンダーを取得する。無ければ作成する",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "正常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          },
          "401": {
            "description": "未ログイン"
          }
        }
      }
    },
    "/api/calendars/{calendar_id}": {
      "get": {
        "summary": "カレンダー基本情報を取得する。参加者本人のみ",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "正常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          },
          "401": {
            "description": "未ログイン"
          },
          "404": {
            "description": "存在しない、または参加していないカレンダー"
          }
        }
      }
    },
    "/api/calendars/{calendar_id}/members": {
      "get": {
        "summary": "カレンダー参加者一覧(スコアランキング)を取得する。参加者本人のみ",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "正常"
          },
          "401": {
            "description": "未ログイン"
          },
          "404": {
            "description": "存在しない、または参加していないカレンダー"
          }
        }
      }
    },
    "/api/calendars": {
      "post": {
        "summary": "グループカレンダーを作成する。作成者は自動的にownerとして参加する",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "作成成功(招待コードを含む)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          },
          "400": {
            "description": "入力エラー"
          },
          "401": {
            "description": "未ログイン"
          }
        }
      }
    },
    "/api/calendars/join": {
      "post": {
        "summary": "招待コードでグループカレンダーに参加する。参加済みなら200(冪等)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarJoin"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "新規に参加した",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          },
          "200": {
            "description": "既に参加済み",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          },
          "400": {
            "description": "入力エラー"
          },
          "401": {
            "description": "未ログイン"
          },
          "404": {
            "description": "招待コードが見つからない"
          }
        }
      }
    }
  },
  "openapi": "3.0.3",
  "components": {
    "schemas": {
...
      "Reaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "user_id": {
            "type": "integer",
            "readOnly": true
          },
          "target_type": {
            "type": "string",
            "enum": [
              "event",
              "post"
            ]
          },
          "target_id": {
            "type": "integer"
          },
          "kind": {
            "type": "string",
            "enum": [
              "fire",
              "thumbs_up",
              "muscle",
              "party"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "kind",
          "target_id",
          "target_type"
        ],
        "additionalProperties": false
      },
      "Calendar": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "readOnly": true
          },
          "type": {
            "type": "string",
            "readOnly": true
          },
          "owner_id": {
            "type": "integer",
            "readOnly": true
          },
          "invite_code": {
            "type": "string",
            "readOnly": true,
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CalendarCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
      "CalendarJoin": {
        "type": "object",
        "properties": {
          "invite_code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          }
        },
        "required": [
          "invite_code"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}